requirements.txt
vpython
pytorch